Move MockCompute to different directory
