+++ webapp/apps/core/views.py
@@ -1,5 +1,3 @@
-import traceback
-import json
 from django.shortcuts import render, get_object_or_404
 from django.http import JsonResponse
